linkstate suite update
Linkstate suite did not use mandatory arguments in update messages and
this caused bug 6373. Now the play.py script is fixed in 44261.
To be able to catch the problem from the bug the TC1_ cases are updated
to use --skipattr with play.py.
Cases TC2_ are to verify linkstate data in the advertised routes.

Change-Id: Ib53725c68595cf507c940ae4cca366b54c16ebaa
Signed-off-by: Peter Gubka <pgubka@cisco.com>