updating bgp clustering test suite due to present bgpcep singleton impl
For Be there must remain the old suites, because they will be too
different from B and C, so differentiating the code inside robot will be
less maintainable and readable

Added configplan code is moved from builder https://git.opendaylight.org/gerrit/45206
and is valid for Be-bgp only. B and C implementations doesnt need it.

Change-Id: I0aa189af2e786abd5ff6b2064c043ef1a8ca28a5
Signed-off-by: Peter Gubka <pgubka@cisco.com>