Update patch set 1
Patch Set 1: Code-Review+1

This looks very good. As additional steps it would be nice to:
a) create a sandbox to see it working in ODL jenkins 
b) test without creating additional interfaces (br-sfc). With those created by the script we should be able to deploy a similar working configuration.

Patch-set: 1
Label: Code-Review=+1