Update patch set 2
Patch Set 2:

Hmm, I think it is time we should change our recommendations. Restconf suite should be named in a way to make it clear that it is not a suite relevant for {project}, and there should be FIXMEs inside to remove this suite when a real functional tests appear.

Patch-set: 2
Reviewer: Gerrit User 2046 <2046@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0