Added 010__sfc_redundancy.robot Test Suite
Some functions used by both Test Suites (Full Deployment and Redundancy)
have been moved to Utils.

The basic TC committed configures a RSP via API Rest and then checks
that the configuration is applied to every Controller (Leader and Followers)

Besides, simplified deployment/json have been used. The nodes configured are
a SF and a SFF. This deployment should be enough to check that flows are
correctly installed, if necessary.

Change-Id: Icb2332131f280d28fafcc380fb380982904bbe70
Signed-off-by: ematejo <jose.angel.mateos.ramos@ericsson.com>