recreating the former test suite for BE stream
the change 45165 fixed the problem for boron stream. But the Be stream
does not pay attention on the presence of mandatory arguments and the updated
test suite with 2 testcases TC1_ and TC2_ is not relevant for Be stream.
Therefor the odl test suite with just one test case is recreated for Be.

Change-Id: I59abf8404684ef3c2bc3ecb30d518a34de97d237
Signed-off-by: Peter Gubka <pgubka@cisco.com>