Update patch set 39
Patch Set 39:

Vratko,

Currently, for the SFC project there is no difference in the configuration files used for beryllium, boron or carbon. The same configuration files are valid for the three releases in the rest-basic suite, as you can see in [0] and [1]. These two sandbox regressions are with this very last commit.
I believe that if there is no compatibility issues between releases there is no point in duplicating configuration files by creating new folders. If the time comes when a new release is not backwards compatible, then we will create it. But at for the time being I believe it is better to keep a single folder.


[0] https://jenkins.opendaylight.org/sandbox/job/sfc-csit-3node-rest-basic-only-beryllium/1/
[1] https://jenkins.opendaylight.org/sandbox/job/sfc-csit-3node-rest-basic-only-boron/

Patch-set: 39
Reviewer: Gerrit User 6322 <6322@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0