Update patch set 4
Patch Set 4: Code-Review+2

> I agree. I didn't want to modify Variables.robot as it's global and
 > only neutron tests are interested in these urls.
 > 
 > In this case we should modify all neutron/__*.robot as well. Do I
 > modify the other robot files here too? Or do I start another
 > review? I think the second approach is better from a git/gerrit
 > point of view.

This patch is good for now.  The Variables.robot is a duplication of the Variables.py file.  We are hoping to move all of these more constant type variables to Variables.robot to keep them in a common location.  But, a variables/neutron.robot file would work too.

let's merge this patch and further improvements can happen if you like.

Patch-set: 4
Label: Code-Review=+2