Update patch set 11
Patch Set 11: Code-Review-1

The only difference between the three robot test files is the COUNT variable. Is there not a way to add the three tests (with 1, 100k and 1M routes) in the same robot file?
Or, if you prefer to have a robot file per scenario, at least avoid code replication. Move the common functions to a different file and import them.

Patch-set: 11