Update patch set 11
Patch Set 11:

> > The only difference between the three robot test files is the
 > COUNT
 > > variable. Is there not a way to add the three tests (with 1, 100k
 > > and 1M routes) in the same robot file?
 > > Or, if you prefer to have a robot file per scenario, at least
 > avoid
 > > code replication. Move the common functions to a different file
 > and
 > > import them.
 > 
 > And how do you want to run 3 suites with just one robot file? 
 > Manually yes, but any job just takes testplan file and runs suites
 > mantioned inside. If you find a way how to parametrize suite via
 > testplan, let me know.

And to andswer your question, it is possible to create a template and run it, so you will have just 3 test cases, for each count one. This solution will provide you much less information on the first look and you will have to "time costly" investigate, what happened. Now it it is much straight forward when comparing several builds.

Patch-set: 11