Update patch set 1
Patch Set 1:

This patch is dependent on the patch submitted for releng/builder[1]. Hopefully, when that patch will be merged than this patch will run successfully.
[1] https://git.opendaylight.org/gerrit/#/c/47234/

Reason for the patch is : 

The link from which builder tries to download the Cassandra package is broken[2][3] which results in the abortion of all Cassandra Jobs.

[2] https://github.com/opendaylight/releng-builder/blob/master/jenkins-scripts/builder.sh#L38
[3] http://apache.osuosl.org/cassandra/2.1.14/apache-cassandra-2.1.14-bin.tar.gz

Patch-set: 1
Reviewer: Gerrit User 4094 <4094@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0