Update patch set 5
Patch Set 5: Code-Review+1

sandbox runs showing this "Wait For Karaf Log" is not broken on the Login:

https://jenkins.opendaylight.org/sandbox/job/topoprocessing-csit-1node-topology-operations-only-boron/7/
https://jenkins.opendaylight.org/sandbox/job/didm-csit-1node-discovery-only-carbon/1/

NOTE: the DIDM job is still failing, but because the actual log message this keyword
is looking for is not there, not because the keyword is broken.  That is a different
issue that will probably need a DIDM expert to check out.

Patch-set: 5
Label: Code-Review=+1