Update patch set 2
Patch Set 2:

> hmm odd, these tests still aren't passing [1]
 > 
 > they work fine in transparent runs [2], so it is obviously SG
 > related, but the rest of CSIT works ok - is anything special done
 > in these tests that isn't done in the rest of CSIT?
 > 
 > [1] https://jenkins.opendaylight.org/sandbox/job/netvirt-vpnservice-csit-1node-openstack-mitaka-learn-boron/3/
 > [2] https://logs.opendaylight.org/releng/jenkins092/netvirt-vpnservice-csit-1node-openstack-mitaka-transparent-carbon/104/archives/log.html.gz

maybe it's a legit bug?  karaf.log is messy and I saw at least one message indicating it could not program a flow.

I wonder if it would be valuable to add a dhcp namespace ping first before trying to ssh to the instances like we do in the other connectivity suites?

Patch-set: 2
Label: Verified=0