Update patch set 11
Patch Set 11:

Hmm, maybe the code would look more pleasant if we relied on TemplatedRequests ability to look for ${folder}.${stream}/ first.
That way we can use ${folder} for openconfig way an ${folder}.boron (and ${folder}.beryllium) for the config subsystem way for doing a particular action.

Patch-set: 11
Label: Verified=0