Update patch set 2
Patch Set 2: Code-Review-1

can you explain why this check is not accurate between the two impls?  the test case looks useful to me.

also, the verify job fails because of a robot.tidy issue. I have a patch to fix that here:
https://git.opendaylight.org/gerrit/48084

once it's merged, I'll rebase this one.

Patch-set: 2