Update patch set 2
Patch Set 2: Code-Review+2

> > can you explain why this check is not accurate between the two
 > > impls?  the test case looks useful to me.
 > >
 > > also, the verify job fails because of a robot.tidy issue. I have
 > a
 > > patch to fix that here:
 > > https://git.opendaylight.org/gerrit/48084
 > >
 > > once it's merged, I'll rebase this one.
 > 
 > 
 > Jamo,
 > 
 > The legacy Netvirt and New Netvirt have differnt bridge
 > configurations. this test looks for is_connected 6 times which
 > suits Legacy Netvirt  as it creates br-int and br-ex. This test is
 > not a common test. We will add a seperate robot file for these
 > checks

ok, merging then.  if we end up with more of these kinds of impl specific differences we can introduce a flag of some sort to allow the same suite to conditionally check things.

one idea would be to have the suite do a karaf feature list -i and if it has odl-ovsdb-openstack in it then we would count 9 ovs managers, else we count 6.  anyway....

Patch-set: 2
Label: Code-Review=+2