updating bgpingest suites to support openconfig for carbon
the support of openconfig in carbon happenes on getting the content
for PUT(or any other rest) messages inside TemplatedRequests

in the csit/variables/bgpuser|bgpingest are already direcotries such as
bgp_peer or bgp_peer.boron. TempletedRequests will take the rest requests
contetnt from appropriate directory. For default directories (applied for
carbon) openconfig is used. For directory.B|Be netconfconnector is used.

Change-Id: I99cd2be0bf94b4234473af5f38e654c5bbaa9a9e
Signed-off-by: Peter Gubka <pgubka@cisco.com>