Update patch set 12
Patch Set 12: Code-Review+1

cool!  I've got nothing to complain about here.

we need to make sure this is ok in the sandbox on both legacy and current openstack jobs.  running it on newton would be good, but not needed for us to merge the patch.  just want to make sure we don't break the good stuff we have running now.

Patch-set: 12
Label: Code-Review=+1