Update patch set 20
Patch Set 20: Code-Review-1

> these jobs had serious failures.  we need to look.  one job seemed
 > to get the VMs spun up
 > and found the IPs like we expect, but the connectivity was not
 > there.  The other job didn't
 > find the IPs like we want.

Interesting.. good thing you ran these.

For the v2 - i'm not sure how well it's supposed to work, but this raised a problem with how we poll - it seems the first VM didn't get DHCP in this case.
What surprised me is that we didn't keep looping on the other VMs to check for their DHCP address, so what could be a single DHCP failure caused all CSIT tests to be IPless and fail.
Would prefer to have polling on all VMs DHCP IP in any case, and then we could optionally SKIP the tests that ping if there is no IP

For the stateful - I thought nothing was working connectivity-wise anyways because of the kernel modules?

Patch-set: 20
Label: Code-Review=-1