Update patch set 1
Patch Set 1:

I'm not sure what benefit we will get here, as the robot report is just going to show pass and not really point out that we have a problem (unless someone is really digging deep in to the logs).  I prefer to let the failures happen and be visible...

but, wont block this patch.

can you move the keyword to the same general vicinity as the other keywords in this library that are dealing with pass/fail bug marking?

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0