Update patch set 20
Patch Set 20:

> Agreed on Alon's comment. Let's get the vm ip loop fixed. Then
 > commit this since learn/stateless and transparent should work. And
 > work on v2 and stateful in parallel.



Here's the problem.  Assume something is totally broken and we want to sit and
spin on every VM (for 180secs) to get an IP address.  Assume they never do. Then
we are wasting 18m in that.

if even 1 VM doesn't get an IP then something is broken.  Are we gaining a lot to
do our best to get the other IPs and see if some are working or are we better off
just trying to get through the test as fast as we can.

Patch-set: 20