remove stale devstack suite setup
the old devstack suite setup is not used any longer and was there to help
bootstrap the devstack gate job which we no longer have.  the keyword was
not used by anything.  The newer "devstack suite setup tests" is what we
are using now and it has been renamed to just "devstack suite setup" which
is more concise

Change-Id: Iedf4ba955ed3544adf6621b5dcf86dfe301ae0ec
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>