Update patch set 18
Patch Set 18:

Jamo, in the single failure we had in last 1node run it seems that the VM did get an IP from DHCP, but the line "^nameserver..." wasn't detected (for getting the DHCP IP). Not sure why.

In any case we didn't really need it for this test case but it now comes with the util.
I'm rerunning now with 1node/3node to see if it persists.

Venkat, for the ARP responder, just need to run the current netvirt 3node job.

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-alonko-upstream-transparent-boron/14/
https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-3node-openstack-mitaka-alonko-upstream-transparent-boron/1/

Patch-set: 18
Label: Verified=0