Update patch set 19
Patch Set 19:

> Venkat, still some issues with new netvirt runs (linked in previous
 > comment)
 > 
 > The 1node external tests are failing for some reason - it seems
 > like the nameserver isn't obtained even though the same VMs are
 > booted.
 > Maybe related to the fact that we don't use the return value of
 > DHCP IPs in external test suite?
 > 
 > The 3node have much more serious failure - the "obtain failed" so
 > we couldn't get the IPs at all. Maybe we're too strict with our
 > grep/regex? Not sure..


Alon,

 The 1node failures probably due to nameserver not getting assigned. So, I have added a new patch to skip extracting the dhcp_ip for the external network tests. Not sure if that is okay!!

The 3node failures are due to the fact that the keyword is waiting long enough until the VM's are assigned the IP Address. the Failing line  is the line to check the "obtained" the lease, which works fine  with the legacy jobs. Maybe there is a  slight delay while installing flow entries in that job?

Patch-set: 19
Label: Verified=0