Update patch set 2
Patch Set 2:

> (1 comment)

Thanks for the comment.

The router-id is a valid parameter in the URL to create a l3vpn, that happens to be optional. I still think that a natural translation to Robot keyword is to have the same keyword, but with this parameter optional as well. This commit is not changing the usage of tests that still want to ignore the optional parameter router-id, thus providing a graceful legacy compatibility.

Providing separate functions because of optional parameters is a pattern that can lead to functions explosions, if you consider the cartesian product of multiple optional arguments.

If you still think so, I can provide a separate function, but I would like to hear more opinions.

Patch-set: 2
Reviewer: Gerrit User 5440 <5440@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0