Update patch set 1
Patch Set 1:

Looking at other users of the function it isn't so bad. Many cases simply grep the output or log the return, so they are not impacted.
In our case, we expected a single value to be returned and used directly so it was an issue. This was just a quick look, so there might be other cases with the same issue,
 but if [1] works, then I think we can make those changes as needed to other tests.

Patch-set: 1