Update patch set 5
Patch Set 5:

> > (1 comment)
 > 
 > yeah that would suck. probably need an upper limit timeout on the
 > whole for loop.
 > 
 > i don't mind waiting to run into it and then fix though, better
 > than the alternative IMO

the other thing this change is going to do is let this specific test
case be marked as PASS even if one or more instances don't get an
IP address.

we need to do this better.

1) when one instance polls for the max time and doesn't get an IP
   then the rest of the VMs are only checked once, not polled against

2) if even one VM did not get an IP we should get a failure in this
   test case.

shall we put a comment TODO in this code and make a trello card, then
merge this, or take the time on this patch to put the better logic?

Patch-set: 5
Label: Verified=0