Make Utils.robot work with more than one session.
Most of the Utils.robot functions were implemented
to use a fixed session variable named 'session'.
This is a problem when having a cluster configured
in which you might want to use some of the tools
with a different controller.
The solution is adding a new argument ${session}
initialized to 'session' to avoid changing the
default behavior of the function.

Change-Id: I8a7341893c77a4d3789522cfe100304ce5658aa5
Signed-off-by: ejisfza <jose.luis.franco.arza@ericsson.com>