Update patch set 1
Patch Set 1: Code-Review-2

> > Can you post some sandbox run for this patch?
 > > Did you push this patch based on some releng observation?
 > 
 > check this out:
 > https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-mitaka-upstream-transparent-carbon/46/archives/log.html.gz#s1-s3-s1-t22-k2-k1-k1-k3-k3
 > 
 > makes me wonder what might be broken, if VMs are not being deleted
 > properly.



Oh, I think I see the problem.  There is a test case above this one (link above)
that deletes that first VM [0] .  so, when we run through the loop of all VMs, that
first VM really is already gone.  The error message in that case is what we 
would expect.  We can merge this patch, but only if we fix the test suite to 
not try to delete an already deleted VM.  otherwise CSIT will report a false 
failure.

[0] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-mitaka-upstream-transparent-carbon/46/archives/log.html.gz#s1-s3-s1-t20

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2