Update patch set 1
Patch Set 1: Code-Review-2

I think something is wrong.

this patch in the sandbox is not able to create the vpns.  my guess is
that something is wrong with the data body being sent with the POST.

here is one that didn't work in the sandbox:
https://logs.opendaylight.org/sandbox/jenkins091/netvirt-csit-1node-openstack-newton-jamo-upstream-transparent-carbon/2/archives/log.html.gz#s1-s4-s1-t15-k6-k4-k1-k3-k2

vs one that did work in releng (obviously without this patch):
https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-mitaka-upstream-transparent-carbon/46/archives/log.html.gz#s1-s4-s1


main thing that stands out to me is that the format of the routes in
the data is not list-like in the sandbox where it fails.  That's as
far as I dug.

Patch-set: 1
Label: Code-Review=-2