Update patch set 1
Patch Set 1:

This is to list provider services, which isn't plumbed in so this will always fail. I am guessing the proper fix would be to have some way for the neutron northbound to query the providers what they support and reply to this api or simply respond to the api.

Patch-set: 1