Update patch set 3
Patch Set 3:

> > man I hate tidy... What is the max line length? Also is there a
 > way
 > > to not have tidy expect a line to be so long? I would really like
 > > to clean this big long line up to make it more readable.
 > 
 > yeah Sam, the line is lengthy, Maybe we can add them to a text file
 > and read the file from robot and process that list. This was
 > discussed in "test" meeting a few weeks back

let's get this patch merged.

then take a look at this one [0] where I moved the module list to a python
variable file.  it should be better for everyone to work with.  it contains
the new modules/changes that are in this patch already.

I'll run it in the sandbox and put a link there.

[0] https://git.opendaylight.org/gerrit/#/c/49501

Patch-set: 3
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0