Update patch set 4
Patch Set 4:

still just working on the refactor, but it's starting to look good with
the logic to fail the test case when not all IPs are found.

the latest patchset #4 is trying to refactor the keywords to make the individual test cases less messy and easier to maintain.

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-transparent-carbon/6/

Patch-set: 4