Update patch set 1
Patch Set 1:

I'm fine with a patch like this as it helps our job to not timeout.  the trouble with long rest requests with only a single node being up is not giving us any failure anyway so it's not helping us track any
issue.

I have a discussion open to figure out what's going on here and maybe a bug/fix will come from it and we can then revert this.

Patch-set: 1
Label: Code-Review=0