Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

I put comments inline, but I'm not sure this is really working. if you look at the console output of the sandbox link [0] you can see that the script is being called, but it returns immediately and none of the debug output (e.g. the echo statements) are not showing up.

not sure why, but probably something simple is causing it to exit silently.

Patch-set: 5
Label: Code-Review=-1