Update patch set 5
Patch Set 5:

> I took a look at random .xml files present in current codebase. I
 > have seen 1 tab, 1 space or 4 spaces indent, never 2 spaces.

even I saw a mix of 1 or 4 spaces, but 2 just seemed to give a better and consistent indent and was the default too.

 > I am starting to think that following style guidelines from diverse
 > projects results in non-homogeneous standard for int/test. I prefer
 > homogeneous look and feel.

following style guidelines was just for the base file, we can modify as it suits to the community.

 > Personally, would prefer indent to be 4 spaces for every file type,
 > just because I am using a relatively dumb editor (mcedit) and my
 > tab key already aligns to 4 spaces regardless of file extension.

sure, but exceptions can be made for .rb, .pp, dockerfiles etc.

Patch-set: 5
Reviewer: Gerrit User 6900 <6900@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0