Fix up CSIT such that extra-route is cleared correctly.
Extra routes should be cleared from a router before an
attempt is made to remove subnets which are the
targets of extra-routes from router.

Also it is better to have Should Contain as the last
line of test scripts as any cleanup after a failure of
Should Contain never gets triggered.

Change-Id: I2987b60afab3d78e8ef3cb84118e72ea44b087b0
Signed-off-by: Vivekanandan Narasimhan <n.vivekanandan@ericsson.com>