Update patch set 6
Patch Set 6:

so, this patch is kind of ahead of the game :) and will not be used/useful until the federation feature is put upstream.

do we want this merged anyway? I'm fine with the patch at this
point, so we can merge and let it be there ready and waiting.

Sandra, you can +1 if you are good.

Patch-set: 6