Update patch set 17
Patch Set 17:

In (1), the 2 test cases failed because of the failure of the keyword Verify VM Received DHCP Lease in start suite.Because of this the variables did not get the correct IP address. 

That same has been fixed by changing the regular expression and has been validated in (2).

So (2) is the final log file of my suite. I just attached (1) to show that the working of Verify VM Received DHCP Lease keyword was not completely accurate. 

Thanks.

Patch-set: 17