Update patch set 1
Patch Set 1: Code-Review+1

This patch is submitted for a bug noticed with the Keywords which use SSHLibrary.Write. When I call the keywords multiple times we see that the SSHLibrary.Write buffer is not flushed, and tries to execute the previous commands. There is not certainty on which command executes, though it is confirmed that the commands received to the procs are in order. This patch will add a SSHLibrary.Read just before the SSHLibrary.Write which makes sure that the buffer is cleaned.

Attaching the passlog: https://logs.opendaylight.org/sandbox/jenkins091/netvirt-csit-1node-openstack-mitaka-upstream-learn-boron-morz/8/archives/log.html.gz

Patch-set: 1
Reviewer: Gerrit User 5779 <5779@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1