Update patch set 1
Patch Set 1:

There are two very similar keywords (difference is whether the prompt is configurable). There could be only one (with prompt argument having a default value).

As for Read, I expect that to result in futile ${timeout} waits, expect for the scenario this is aimed to fix, which is caused by some other keyword not waiting for its prompt. The other keyword should be fixed instead of adding Read to these keywords. In either case, Sandbox would be useful to see what is happening.

Patch-set: 1