Update patch set 1
Patch Set 1:

> So far I see no place where the added Read fetches anything. Also,
 > looking at RelEng results, The failures come from a failed ping
 > [0], but the previous commands seem to be functioning normally [1].
 > 
 > [0] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-mitaka-upstream-learn-boron/118/archives/log.html.gz#s1-s4-s1-t9-k10-k13-k2-k4-k1-k2
 > [1] https://logs.opendaylight.org/releng/jenkins092/netvirt-csit-1node-openstack-mitaka-upstream-learn-boron/118/archives/log.html.gz#s1-s4-s1-t9-k10-k12-k1-k3

Hi Vratko,

Agree with you, that i normal scenarios even we did not face any problems with it.
But when we started scripting for the HWVTEP testcases, we started noticing this error. There are many configurations commands which needed to be executed in the TOOLS_SYSTEM where we observed this.

Unfortunately I don't have a passlog which shows this problem. I would give you a failure log to demonstrate this issue next week. With the extra Read, it started working normal, without it, it fails.

Thanks,
Abbas

Patch-set: 1