Update patch set 4
Patch Set 4:

the idea so far is to call this distcompare tool from our distribution
job which creates the distribution to put on nexus. It would be up to
the bash scripts calling this tool to set up the distros (download,
unzip, move to /tmp/, etc). Hopefully that is the only work required.

the output would then end up in the distribution jobs console log.

I suppose creating a file with the distro diffs and put on nexus could
be a future enhancement so when we go to nexus and see the distro in
question, we can easily see the diffs and not have to track down the
actual distribution job that created it.

Patch-set: 4
Label: Verified=0