Update patch set 5
Patch Set 5:

So I guess a good middle ground is to use 4 spaces if we don't have a clear "normal" external style. Puppet, Ruby and Gemfiles for example seem standardized on 2 spaces. Sphynx clearly specifies 2 spaces for rst. There don't seem to be clear styles for Vagrantfiles and Dockerfiles, so nbd to use 4. Not sure about XML. Everything else is already 4.

Patch-set: 5