Update patch set 2
Patch Set 2: Code-Review+1

Hi Venkat, thank you for reviewing the patch!

Could you make sure that if renaming is correct or not, and please review the patch again?

I think renaming is correct.
I don't think the new NetVirt uses the scripts for the JVM plots which are written in these scriptplans. The new NetVirt CSIT jobs don't install features which are required for the JVM plots. The current files are not needed for the new NetVirt jobs as of now.

In my understanding, the current files were created for the legacy NetVirt CSIT jobs. Therefore, the names of these files should have be renamed with the following patch.

https://git.opendaylight.org/gerrit/#/c/47859/

However, unfortunately, I think, the above patch have forgotten renaming the names of these files.

To track the history of the changes of the scriptplans for the legacy NetVirt, I think renaming is the right choice here.

Among the testplan files for new NetVirt, only the file "csit/testplan/netvirt-3node-openstack.txt" has the following line.

./netvirt-legacy-clustering.txt:integration/test/csit/suites/integration/Create_JVM_Plots.robot

It may give us an impression that the new NetVirt uses the JVM plots, but, again, the new NetVirt CSIT job don't install features for the JVM plots. So I believe that they don't use JVM plots. It seems to me that the above line was mistakenly left by the change of the above patch (Gerrit 47859).

And, I think keeping unused files are not good. So if these files are not used, we should not keep them here. Otherwise, other people get confused on these unused files. They would try to get answers about how these files are used, however, there is't answers so they would just waste their time.

If the current files are really used by the new NetVirt jobs, I'll updated the patch to copy the files instead of renaming.

Patch-set: 2
Reviewer: Gerrit User 396 <396@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1