Update patch set 42
Patch Set 42:

Mor, can you quickly create a report with a failing DHCP check? (I'm pretty sure there was one already, but just for Alon to see).

We really should get this patch in already - the 3node jobs are failing on what looks like DHCP issues - without this patch it's really hard to tell whether those are real regression or just a test issue - so I'd be happy if we can get this in asap

Patch-set: 42