Update patch set 6
Patch Set 6: Code-Review+2

Guy,

I think the apathy here is that we know this change has no real affect on 
anything running upstream. The code is just a NOOP with what we have currently.

I have merged it, as the assumption is federation features are coming soon.

JamO

Patch-set: 6
Label: Code-Review=+2