Update patch set 7
Patch Set 7:

(1 comment)

Jamo, pushed a patch that reworked the file a bit. It works now as expected so you can iterate from there.

Question: the diff just takes what is in new - it compares new against the old, but only copies new gerrits to the diff list. Would that always be right to assume new is a progression from old? What about a case where there was a reverted patch - guess that would be a new patch so the assumption would hold.

Patch-set: 7