Update patch set 7
Patch Set 7:

(6 comments)

> (1 comment)
 > 
 > Jamo, pushed a patch that reworked the file a bit. It works now as
 > expected so you can iterate from there.
 > 
 > Question: the diff just takes what is in new - it compares new
 > against the old, but only copies new gerrits to the diff list.
 > Would that always be right to assume new is a progression from old?
 > What about a case where there was a reverted patch - guess that
 > would be a new patch so the assumption would hold.

the way I *think* it's working is that we have a *newer* distro with it's list of most recent patches that were merged. The *older* distro has it's list. Any patch that is only in the *new* list will be 
considered a difference. so, a revert that is merged will just show
up as a patch that is there in *new* but not *old*

reasonable?

Patch-set: 7
Label: Verified=0