Update patch set 17
Patch Set 17: Code-Review-1

Mor,

wondering if this might be a more cleaner idea (test robot code below you can fool with locally if you want)

make a [Tags] line for this specific tempest test with something like:
[Tags]    skip_if_stable/mitaka

Then in our tempest keyword that runs the test, we just need one line at the beginning to
Return From Keyword If    "skip_if_${OPENSTACK_BRANCH}" in @{TEST_TAGS}

that's the python "in" conditional there, so we can have any number of tags for the test case,
and it should still work.

seems cleaner/more-clear to me that way, but just a suggestion.


*** Settings ***
*** Variables ***
${openstack_branch}    stable/mitaka

*** Test Cases ***
Skiping it or not
  [Tags]  skip_if_stable/mitaka
  Log To Console    before KW
  Run The Test
  Log To Console    after KW

Skiping it or not
  [Tags]  skip_if_stable/newton
  Log To Console    before KW
  Run The Test
  Log To Console    after KW

*** Keywords ***
Run The Test
  Log To Console    Inside our local keyword
  Return From Keyword If    "skip_if_${openstack_branch}" in @{TEST_TAGS}
  Log To Console    Still Inside our local keyword

Patch-set: 17
Label: Code-Review=-1