Update patch set 19
Patch Set 19: Code-Review-1

> > Mor,
 > >
 > > wondering if this might be a more cleaner idea (test robot code
 > > below you can fool with locally if you want)
 > >
 > > make a [Tags] line for this specific tempest test with something
 > > like:
 > > [Tags]    skip_if_stable/mitaka
 > >
 > > Then in our tempest keyword that runs the test, we just need one
 > > line at the beginning to
 > > Return From Keyword If    "skip_if_${OPENSTACK_BRANCH}" in
 > > @{TEST_TAGS}
 > >
 > > that's the python "in" conditional there, so we can have any
 > number
 > > of tags for the test case,
 > > and it should still work.
 > >
 > > seems cleaner/more-clear to me that way, but just a suggestion.
 > >
 > >
 > > *** Settings ***
 > > *** Variables ***
 > > ${openstack_branch}    stable/mitaka
 > >
 > > *** Test Cases ***
 > > Skiping it or not
 > > [Tags]  skip_if_stable/mitaka
 > > Log To Console    before KW
 > > Run The Test
 > > Log To Console    after KW
 > >
 > > Skiping it or not
 > > [Tags]  skip_if_stable/newton
 > > Log To Console    before KW
 > > Run The Test
 > > Log To Console    after KW
 > >
 > > *** Keywords ***
 > > Run The Test
 > > Log To Console    Inside our local keyword
 > > Return From Keyword If    "skip_if_${openstack_branch}" in
 > > @{TEST_TAGS}
 > > Log To Console    Still Inside our local keyword
 > 
 > I don't understand how to use the tags, how I add tag to a specipic
 > test? and how I get @{TEST_TAGS} var in DevstackUtils.robot?
 > thanks!

yeah, we don't use Tags very much in OpenDaylight Robot suites. they can
be pretty useful though.

so, your test case would be like this:

tempest.scenario.test_network_basic_ops.TestNetworkBasicOps.test_hotplug_nic
    [Tags]    skip_if_stable/mitaka
    ${TEST_NAME}    ${exclusion_regex}    ${tempest_config_file}

Then your library to run the tempest stuff would have no change except one new
line at the beginning.

    Return From Keyword If    "skip_if_${openstack_branch}" in @{TEST_TAGS}



that @{TEST_TAGS} is given to us for free by robot framework itself. It
could be None, but if [Tags] are given it will hold whatever those tags
are.

Oh, to get ${OPENSTACK_BRANCH} I think we have a little work to do on the
releng/builder side. Let me try to put a patch for that. It should be
something we are passing in on the robot command line which will make
it available (and will override that ${OPENSTACK_BRANCH} that is defaulted
to stable/liberty in the lib file)

Patch-set: 19
Label: Code-Review=-1