Update patch set 22
Patch Set 22: Code-Review+1

ok, finally I have something to show. with this latest patch set
22, you can see how the test case is skipped (and marked pass)

https://logs.opendaylight.org/sandbox/jenkins091/netvirt-csit-1node-openstack-mitaka-jamo-upstream-stateful-boron/15/archives/log.html.gz#s1-s2-s1-t3-k2-k1

the rest of the job is failing all over the place and I'm wondering
if we might have something broken in netvirt all of a sudden. but,
that's besides the point.

job 17 is still running, but it's only running tempest and maybe
will be a cleaner look:

https://jenkins.opendaylight.org/sandbox/job/netvirt-csit-1node-openstack-mitaka-jamo-upstream-stateful-boron/17/

Patch-set: 22
Label: Code-Review=+1