Update patch set 2
Patch Set 2: Code-Review-1

we can put this patch if it's deemed neccessary, but I think we really
should view this sudden set of failures as a true regression that
we can't let in to Boron. Until we know what causes it, I'd rather
not work around it and just let things fail.

Patch-set: 2
Label: Code-Review=-1