Update patch set 1
Patch Set 1: Code-Review+2

> > I think it is better to add tempest tests to a differnt suite or
 > at
 > > the end to avoid such confusion.
 > 
 > Venkat, what do you mean?
 > Tempest is in its own suite (tempest), and this patch moves it to
 > be after the connectivity tests..

I think Venkat wants tempest to have it's own job unrelated to the
other suites here. One concern I think was that new tempest cases
might be coming in and creating failures (either in tempest itself
or other cases later in the suite).

Personally, even though it adds some headaches, I prefer to keep as
many functional tests in the same live setup as possible. As long
as they are not giving us false failures, I think we have better
chances to flush out more bugs. I get the sense this isn't everyone's
opinion though.

Let me merge this patch as is, since it's not really changing the
overall purpose/coverage of our jobs. But, if we want to discuss a
new job just for tempest, we can bring it to the mailing list and 
see what others think.

Patch-set: 1
Label: Code-Review=+2